Refactor missing background loop into reusable worker
